requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas
scikit-learn
tqdm
transformers
pillow
matplotlib
numpy
